Added slicer for selecting start and end date
A slicer has been added which allows the user to filter the data between two dates. It currently allows the user to choose data from the last 30 days since this is coded into the SQL queries for FormatedTimestamp. 

A column called DateOnlySwedish was added in each of the four tables that we are currently working on which extracts only the date from the SwedishTime column. In order for this to work, the column was created with DAX and then changed with built in Power BI functions to have the data format "Date" and the date format to be "Short Date".

To make the connections work, DAX was used to create DateTable that takes the earliest date from the union of each table (power, pressure, temperature, motor), as well as todays date, and sets them as start and end date. This is then connected with a one-to-many connection to DateOnlySwedish of each table.